# docs/swagger.yaml
basePath: /v1
definitions:
  models.APIError:
    properties:
      error_message:
        type: string
      error_status:
        type: integer
    type: object
  models.Address:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      domain:
        type: string
      expires_at:
        type: string
      first_seen:
        type: string
      group:
        $ref: '#/definitions/models.Group'
      group_id:
        type: integer
      host_fqdn:
        type: string
      hostname:
        type: string
      id:
        type: integer
      ilo_api_flavour:
        type: string
      ilo_fqdn:
        type: string
      ilo_ip:
        type: string
      ilo_password:
        type: string
      ilo_port:
        type: string
      ilo_user:
        type: string
      ip:
        type: string
      ks:
        type: string
      last_seen:
        type: string
      last_seen_relay:
        description: DHCP parameters
        type: string
      mac:
        type: string
      missing_options:
        type: string
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
      updated_at:
        type: string
    type: object
...
  models.DeviceClass:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      vendor_class:
        type: string
    type: object
  models.DeviceClassForm:
    properties:
      name:
        type: string
      vendor_class:
        type: string
    type: object
  models.Group:
    properties:
      address:
        items:
          $ref: '#/definitions/models.Address'
        type: array
      bootdisk:
        type: string
      callbackurl:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      dns:
        type: string
      id:
        type: integer
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      ntp:
        type: string
      option:
        items:
          $ref: '#/definitions/models.Option'
        type: array
      options:
        additionalProperties:
          type: string
        type: object
      password:
        type: string
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      syslog:
        type: string
      updated_at:
        type: string
      vlan:
        type: string
    type: object
  models.GroupForm:
    properties:
      bootdisk:
        type: string
      callbackurl:
        type: string
      dns:
        type: string
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      ntp:
        type: string
      options:
        additionalProperties:
          type: string
        type: object
      password:
        type: string
      pool_id:
        type: integer
      syslog:
        type: string
      vlan:
        type: string
    type: object
...
  models.Option:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      address_id:
        type: integer
      created_at:
        type: string
      data:
        type: string
      deleted_at:
        type: string
      device_class_id:
        type: integer
      id:
        type: integer
      opcode:
        type: integer
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      priority:
        type: integer
      updated_at:
        type: string
    required:
    - data
    - opcode
    - priority
    type: object
...
  models.Pool:
    properties:
      authorized_vlan:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      end_address:
        type: string
      gateway:
        type: string
      id:
        type: integer
      lease_time:
        type: integer
      managed_reference:
        type: string
      name:
        type: string
      net_address:
        type: string
      netmask:
        type: integer
      only_serve_reimage:
        type: boolean
      start_address:
        type: string
      updated_at:
        type: string
    required:
    - end_address
    - gateway
    - lease_time
    - name
    - netmask
    - start_address
    type: object
...
info:
  contact: {}
  description: VMware Imaging Appliances written in GO with full HTTP-REST
  title: go-via
  version: "0.1"
paths:
...
  /device_classes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceClass'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get all device classes
      tags:
      - device_classes
    post:
      consumes:
      - application/json
      parameters:
      - description: Add an device class
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.DeviceClassForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create a new device class
      tags:
      - device_classes
  /device_classes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Remove an existing device class
      tags:
      - device_classes
    get:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get an existing device class
      tags:
      - device_classes
    patch:
      consumes:
      - application/json
      parameters:
      - description: DeviceClass ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an ip device class
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.DeviceClassForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update an existing device class
      tags:
      - device_classes
  /device_classes/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fields to search for
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.DeviceClass'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceClass'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search for an device class
      tags:
      - device_classes
  /groups:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Add ip group
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.GroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Remove an existing group
      tags:
      - groups
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get an existing group
      tags:
      - groups
    patch:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an group
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.GroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update an existing group
      tags:
      - groups
...
swagger: "2.0"